# api/openapi.json
{
  "components": {
    "schemas": {
      "Agent": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "provider_id": {
            "type": "string"
          },
          "revoked_at": {
            "format": "int64",
            "type": "integer"
          },
          "risk_class": {
            "format": "int64",
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner",
          "provider_id",
          "version",
          "status",
          "risk_class",
          "created_at",
          "revoked_at"
        ],
        "type": "object"
      },
...
      "AuthorizationRequest": {
        "additionalProperties": false,
        "properties": {
          "transaction": {
            "$ref": "#/components/schemas/TransactionV1"
          }
        },
        "required": [
          "transaction"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "type": "object"
      },
...
      "Response": {
        "additionalProperties": false,
        "properties": {
          "challenge_id": {
            "type": "string"
          },
          "grant": {
            "$ref": "#/components/schemas/Grant"
          },
          "policy": {
            "$ref": "#/components/schemas/Result"
          }
        },
        "required": [
          "policy"
        ],
        "type": "object"
      },
...
      "RiskRequest": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "challenge_id": {
            "type": "string"
          },
          "mandate_id": {
            "type": "string"
          },
          "transaction_hash": {
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "mandate_id",
          "challenge_id",
          "transaction_hash"
        ],
        "type": "object"
      },
...
      "TransactionV1": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "type": "string"
          },
          "agent_id": {
            "type": "string"
          },
          "amount": {
            "format": "int64",
            "type": "integer"
          },
          "category": {
            "type": "string"
          },
          "condition": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "fees": {
            "format": "int64",
            "type": "integer"
          },
          "mandate_id": {
            "type": "string"
          },
          "merchant_id": {
            "type": "string"
          },
          "offer_id": {
            "type": "string"
          },
          "offer_revision": {
            "format": "int64",
            "type": "integer"
          },
          "product": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "quantity": {
            "format": "int64",
            "type": "integer"
          },
          "schema_version": {
            "format": "int64",
            "type": "integer"
          },
          "shipping": {
            "format": "int64",
            "type": "integer"
          },
          "unit_amount": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "schema_version",
          "mandate_id",
          "agent_id",
          "merchant_id",
          "offer_id",
          "offer_revision",
          "action",
          "purpose",
          "product",
          "category",
          "condition",
          "quantity",
          "unit_amount",
          "fees",
          "shipping",
          "amount",
          "currency"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearer": {
        "description": "Sandbox owner or registered agent credential",
        "scheme": "bearer",
        "type": "http"
      },
      "callbackHMAC": {
        "description": "Hex HMAC-SHA256 over purpose prefix plus canonical Callback JSON",
        "in": "header",
        "name": "X-Sandbox-Signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Amounts are integer minor units (RUB kopecks). Timestamps are UTC Unix seconds. Owner and agent credentials are distinct. Sandbox adapter only; production OIDC is a replaceable authentication boundary. No real funds. Idempotency keys are scoped by principal+operation; canonical typed JSON is compared and changed bodies return 409. GET current state after an idempotent retry because the stored response may say PENDING.",
    "title": "NORTH sandbox authorization API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agents": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[!-~]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {},
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "agent": {
                      "$ref": "#/components/schemas/Agent"
                    },
                    "agent_id": {
                      "type": "string"
                    },
                    "agent_token": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "agent_id",
                    "agent_token",
                    "agent"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful result; payment/refund commands may still be PENDING"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner: register sandbox agent and receive credential once (retry returns same credential)"
      }
    },
    "/v1/agents/revoke": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[!-~]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "agent_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "agent_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful result; payment/refund commands may still be PENDING"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner: revoke agent"
      }
    },
    "/v1/authorizations": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[!-~]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthorizationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            },
            "description": "Successful result; payment/refund commands may still be PENDING"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Agent: evaluate proposal; return ALLOW grant, DENY trace or ASK_USER challenge"
      }
    },
    "/v1/challenges/approve": {
      "post": {
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "pattern": "^[!-~]+$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "status"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Successful result; payment/refund commands may still be PENDING"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Safe structured error; see error.code"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner: approve exact risk challenge; then agent retries with a new idempotency key"
      }
    },
...
  }
}